added webpack to bundle js files
