+++ html/board.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="en" ng-app="cardApp">
+
 <head>
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta charset="utf-8">
@@ -9,6 +10,7 @@
   <link rel="stylesheet" href="../css/ribbon.css">
   <script type="text/javascript" src="/../node_modules/angular/angular.min.js"></script>
 </head>
+
 <body>
 
   <div ng-controller="MainCtrl as vm">
@@ -54,7 +56,7 @@
             </div>
             <hr/>
             <br/>
-            <div class="row" >
+            <div class="row">
               <div class="flex">
                 <div class="card-list my-board">
                   <div ng-repeat="card in vm.myBoard">
@@ -67,21 +69,24 @@
 
           <div class="hand">
             <div class="cards hand">
-              <card tabindex="2" ng-repeat="card in vm.player.HandCards" ng-click="vm.isItMyTurn && vm.addCardTo(card, vm.myBoard, vm.player.HandCards, card.Cost, vm.player.Mana)" card-info="card" player-mana="vm.player.Mana"></card>
+              <card tabindex="2" ng-repeat="card in vm.player.HandCards" ng-click="vm.isItMyTurn && vm.addCardTo(card, vm.myBoard, vm.player.HandCards, card.Cost, vm.player.Mana)"
+                card-info="card" player-mana="vm.player.Mana"></card>
             </div>
           </div>
 
           <div class="player-stats" ng-click="vm.setActiveTarget(vm.player)" ng-class="{'target':vm.activeTarget === vm.player}">
             <div class="stat">
               <svg style="width:32px;height:32px" viewBox="0 0 24 24">
-                <path fill="#fff" d="M12,2A10,10 0 0,1 22,12A10,10 0 0,1 12,22A10,10 0 0,1 2,12A10,10 0 0,1 12,2M12,4A8,8 0 0,0 4,12A8,8 0 0,0 12,20A8,8 0 0,0 20,12A8,8 0 0,0 12,4M11,17V16H9V14H13V13H10A1,1 0 0,1 9,12V9A1,1 0 0,1 10,8H11V7H13V8H15V10H11V11H14A1,1 0 0,1 15,12V15A1,1 0 0,1 14,16H13V17H11Z" />
+                <path fill="#fff" d="M12,2A10,10 0 0,1 22,12A10,10 0 0,1 12,22A10,10 0 0,1 2,12A10,10 0 0,1 12,2M12,4A8,8 0 0,0 4,12A8,8 0 0,0 12,20A8,8 0 0,0 20,12A8,8 0 0,0 12,4M11,17V16H9V14H13V13H10A1,1 0 0,1 9,12V9A1,1 0 0,1 10,8H11V7H13V8H15V10H11V11H14A1,1 0 0,1 15,12V15A1,1 0 0,1 14,16H13V17H11Z"
+                />
               </svg>
               : {{vm.player.Mana}}
             </div>
             <img src="../img/card.png" class="player"></img>
             <div class="stat">
               <svg style="width:32px;height:32px" viewBox="0 0 24 24">
-                <path fill="#fff" d="M12,21.35L10.55,20.03C5.4,15.36 2,12.27 2,8.5C2,5.41 4.42,3 7.5,3C9.24,3 10.91,3.81 12,5.08C13.09,3.81 14.76,3 16.5,3C19.58,3 22,5.41 22,8.5C22,12.27 18.6,15.36 13.45,20.03L12,21.35Z" />
+                <path fill="#fff" d="M12,21.35L10.55,20.03C5.4,15.36 2,12.27 2,8.5C2,5.41 4.42,3 7.5,3C9.24,3 10.91,3.81 12,5.08C13.09,3.81 14.76,3 16.5,3C19.58,3 22,5.41 22,8.5C22,12.27 18.6,15.36 13.45,20.03L12,21.35Z"
+                />
               </svg>
               : {{vm.player.Hp}}
             </div>
@@ -90,14 +95,16 @@
 
             <div class="stat">
               <svg style="width:32px;height:32px" viewBox="0 0 24 24">
-                <path fill="#fff" d="M12,2A10,10 0 0,1 22,12A10,10 0 0,1 12,22A10,10 0 0,1 2,12A10,10 0 0,1 12,2M12,4A8,8 0 0,0 4,12A8,8 0 0,0 12,20A8,8 0 0,0 20,12A8,8 0 0,0 12,4M11,17V16H9V14H13V13H10A1,1 0 0,1 9,12V9A1,1 0 0,1 10,8H11V7H13V8H15V10H11V11H14A1,1 0 0,1 15,12V15A1,1 0 0,1 14,16H13V17H11Z" />
+                <path fill="#fff" d="M12,2A10,10 0 0,1 22,12A10,10 0 0,1 12,22A10,10 0 0,1 2,12A10,10 0 0,1 12,2M12,4A8,8 0 0,0 4,12A8,8 0 0,0 12,20A8,8 0 0,0 20,12A8,8 0 0,0 12,4M11,17V16H9V14H13V13H10A1,1 0 0,1 9,12V9A1,1 0 0,1 10,8H11V7H13V8H15V10H11V11H14A1,1 0 0,1 15,12V15A1,1 0 0,1 14,16H13V17H11Z"
+                />
               </svg>
               : {{vm.enemy.Mana}}
             </div>
             <img src="../img/card.png" class="player"></img>
             <div class="stat">
               <svg style="width:32px;height:32px" viewBox="0 0 24 24">
-                <path fill="#fff" d="M12,21.35L10.55,20.03C5.4,15.36 2,12.27 2,8.5C2,5.41 4.42,3 7.5,3C9.24,3 10.91,3.81 12,5.08C13.09,3.81 14.76,3 16.5,3C19.58,3 22,5.41 22,8.5C22,12.27 18.6,15.36 13.45,20.03L12,21.35Z" />
+                <path fill="#fff" d="M12,21.35L10.55,20.03C5.4,15.36 2,12.27 2,8.5C2,5.41 4.42,3 7.5,3C9.24,3 10.91,3.81 12,5.08C13.09,3.81 14.76,3 16.5,3C19.58,3 22,5.41 22,8.5C22,12.27 18.6,15.36 13.45,20.03L12,21.35Z"
+                />
               </svg>
               : {{vm.enemy.Hp}}
             </div>
@@ -122,7 +129,7 @@
     </div>
   </div>
 </body>
-<script type="text/javascript" src="../js/app.js"></script>
+<script type="text/javascript" src="../dist/app.bundle.js"></script>
 <script type="text/javascript" src="../js/components/card.js"></script>
 
 </html>
