+++ html/index.html
@@ -1,11 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta charset="utf-8">
   <title></title>
   <link rel="stylesheet" href="../css/style.css">
 </head>
+
 <body style="overflow-y:hidden;">
 
   <a href="board.html">
@@ -16,4 +18,5 @@
   </a>
 
 </body>
+
 </html>
